+++ ASDKSample/ASDKSample/StageTestData.swift
@@ -17,13 +17,14 @@
 //  limitations under the License.
 //
 
-import Foundation
+import TinkoffASDKCore
+import TinkoffASDKUI
 
 /// Тестовые данные для проведения тестовых платежей
 public enum StageTestData {
     /// Открытый ключ для шифрования карточных данных (номер карты, срок дейсвия и секретный код)
     // swiftlint:disable:next line_length
-    public static let testPublicKey = "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAqBiorLS9OrFPezixO5lSsF+HiZPFQWDO7x8gBJp4m86Wwz7ePNE8ZV4sUAZBqphdqSpXkybM4CJwxdj5R5q9+RHsb1dbMjThTXniwPpJdw4WKqG5/cLDrPGJY9NnPifBhA/MthASzoB+60+jCwkFmf8xEE9rZdoJUc2p9FL4wxKQPOuxCqL2iWOxAO8pxJBAxFojioVu422RWaQvoOMuZzhqUEpxA9T62lN8t3jj9QfHXaL4Ht8kRaa2JlaURtPJB5iBM+4pBDnqObNS5NFcXOxloZX4+M8zXaFh70jqWfiCzjyhaFg3rTPE2ClseOdS7DLwfB2kNP3K0GuPuLzsMwIDAQAB"
+    public static let testPublicKey = "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA5Yg3RyEkszggDVMDHCAGzJm0mYpYT53BpasrsKdby8iaWJVACj8ueR0Wj3Tu2BY64HdIoZFvG0v7UqSFztE/zUvnznbXVYguaUcnRdwao9gLUQO2I/097SHF9r++BYI0t6EtbbcWbfi755A1EWfu9tdZYXTrwkqgU9ok2UIZCPZ4evVDEzDCKH6ArphVc4+iKFrzdwbFBmPmwi5Xd6CB9Na2kRoPYBHePGzGgYmtKgKMNs+6rdv5v9VB3k7CS/lSIH4p74/OPRjyryo6Q7NbL+evz0+s60Qz5gbBRGfqCA57lUiB3hfXQZq5/q1YkABOHf9cR6Ov5nTRSOnjORgPjwIDAQAB"
 
     /// Уникальный идентификатор терминала, выдается Продавцу Банком на каждый магазин.
     public static let terminalKey = "TestSDK"
@@ -33,4 +34,29 @@ public enum StageTestData {
 
     /// Идентификатор клиента в системе продавца. Например, для этого идентификатора будут сохраняться список карт.
     public static let customerKey = "TestSDK_CustomerKey1"
+}
+
+struct SdkCredentials: Equatable, Codable {
+    let uuid: String
+    let name: String
+    let description: String
+
+    // Actual sdk Data
+    let publicKey: String
+    let terminalKey: String
+    let terminalPassword: String
+    let customerKey: String
+}
+
+extension SdkCredentials {
+
+    static let test = Self(
+        uuid: UUID().uuidString,
+        name: "Test",
+        description: "Test stage",
+        publicKey: StageTestData.testPublicKey,
+        terminalKey: StageTestData.terminalKey,
+        terminalPassword: StageTestData.terminalPassword,
+        customerKey: StageTestData.customerKey
+    )
 }